Convert `--local-executor` in an integer flag, defaults to 1 i.e. enabled, use 0 to disable

+++ proxy/proxy.py
@@ -13,13 +13,14 @@ import sys
 import time
 import logging
 
+from pathlib import Path
 from typing import List, Optional, Any
 
 from .core.acceptor import AcceptorPool, ThreadlessPool, Listener
 from .core.event import EventManager
 from .common.utils import bytes_
 from .common.flag import FlagParser, flags
-from .common.constants import DEFAULT_LOG_FILE, DEFAULT_LOG_FORMAT, DEFAULT_LOG_LEVEL
+from .common.constants import DEFAULT_LOCAL_EXECUTOR, DEFAULT_LOG_FILE, DEFAULT_LOG_FORMAT, DEFAULT_LOG_LEVEL
 from .common.constants import DEFAULT_OPEN_FILE_LIMIT, DEFAULT_PLUGINS, DEFAULT_VERSION
 from .common.constants import DEFAULT_ENABLE_DASHBOARD, DEFAULT_WORK_KLASS, DEFAULT_PID_FILE
 
@@ -206,8 +207,12 @@ class Proxy:
 
     def _write_pid_file(self) -> None:
         if self.flags.pid_file is not None:
-            # NOTE: Multiple instances of proxy.py running on
-            # same host machine will currently result in overwriting the PID file
+            if Path(self.flags.pid_file).exists():
+                raise ValueError(
+                    'pid file {0} already exists'.format(
+                        self.flags.pid_file,
+                    ),
+                )
             with open(self.flags.pid_file, 'wb') as pid_file:
                 pid_file.write(bytes_(os.getpid()))
 
@@ -217,7 +222,7 @@ class Proxy:
 
     @property
     def remote_executors_enabled(self) -> bool:
-        return self.flags.threadless and not self.flags.local_executor
+        return self.flags.threadless and self.flags.local_executor != int(DEFAULT_LOCAL_EXECUTOR)
 
 
 def main(**opts: Any) -> None:
